refactor: SessionProperty 추가
