+++ src/main/java/nextstep/session/domain/property/SessionStatus.java
@@ -1,9 +1,13 @@
-package nextstep.session.domain;
+package nextstep.session.domain.property;
 
 public enum SessionStatus {
     PREPARING, ENROLLING, CLOSED;
 
     public boolean canEnroll() {
         return this == ENROLLING;
+    }
+
+    public boolean canNotEnroll() {
+        return !canEnroll();
     }
 }
