+++ src/test/java/nextstep/session/domain/SessionTest.java
@@ -6,7 +6,8 @@ import nextstep.session.domain.image.SessionCoverImageType;
 import nextstep.session.domain.payment.SessionCapacity;
 import nextstep.session.domain.payment.SessionFee;
 import nextstep.session.domain.payment.SessionPayments;
-import nextstep.session.domain.payment.SessionType;
+import nextstep.session.domain.property.SessionStatus;
+import nextstep.session.domain.property.SessionType;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
