+++ src/test/java/nextstep/session/domain/property/SessionPropertyTest.java
@@ -1,0 +1,15 @@
+package nextstep.session.domain.property;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class SessionPropertyTest {
+
+    @DisplayName("SessionProperty 인스턴스 생성")
+    @Test
+    public void testConstructor() {
+        assertDoesNotThrow(() -> new SessionProperty(SessionStatus.ENROLLING, SessionType.FREE));
+    }
+}
